Extracting class to a better home
These classes were inextricably intertwined as a matter of convenience.
Now they have moved to be more readily reused.
