+++ app/presenters/sipity/controllers/visitors/core/work_area_presenter.rb
@@ -60,81 +60,6 @@ module Sipity
           end
 
           attr_accessor :submission_window, :start_a_submission_action
-
-          # Responsible for rendering the form for filtering out the various
-          # processing actions.
-          class FilterFormPresenter < Curly::Presenter
-            presents :work_area
-
-            def select_tag_for_processing_state
-              select_tag(
-                work_area.input_name_for_select_processing_state,
-                options_from_collection_for_select(
-                  work_area.processing_states_for_select, :to_s, :humanize, work_area.processing_state
-                ), include_blank: true, class: 'form-control'
-              ).html_safe
-            end
-
-            def select_tag_for_sort_order
-              select_tag(
-                work_area.input_name_for_select_sort_order,
-                options_from_collection_for_select(
-                  work_area.order_options_for_select, :to_s, :humanize, work_area.order
-                ), include_blank: true, class: 'form-control'
-              ).html_safe
-            end
-
-            def submit_button(dom_class: 'btn btn-default', name: 'Filter')
-              submit_tag(name, class: dom_class)
-            end
-
-            private
-
-            attr_reader :work_area
-          end
-
-          # Responsible for rendering a given work within the context of the Dashboard.
-          #
-          # @note This could be extracted outside of this namespace
-          class WorkPresenter < Curly::Presenter
-            presents :work
-
-            def path
-              PowerConverter.convert(work, to: :access_path)
-            end
-
-            def work_type
-              work.work_type.to_s.humanize
-            end
-
-            def creator_names_to_sentence
-              creators.to_sentence
-            end
-
-            def program_names_to_sentence
-              Array.wrap(repository.work_attribute_values_for(work: work, key: 'program_name')).to_sentence
-            end
-
-            def date_created
-              work.created_at.strftime('%a, %d %b %Y')
-            end
-
-            def processing_state
-              work.processing_state.to_s.humanize
-            end
-
-            def title
-              work.title.to_s.html_safe
-            end
-
-            private
-
-            attr_reader :work
-            def creators
-              # The repository comes from the underlying context; Which is likely a controller.
-              @creators ||= Array.wrap(repository.scope_users_for_entity_and_roles(entity: work, roles: Models::Role::CREATING_USER))
-            end
-          end
         end
       end
     end
